==> src/02_indicator_introductions_per_year.Rmd <==
---
title: "Indicator: Number of new introductions of alien species per year"
author:
- Damiano Oldoni
- Yasmine Verzelen
- Stijn Van Hoey
- Tim Adriaens
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
---

This document describes how to build indicators of Invasive Alien Species based 
on checklist data. In particular, this document takes into account the number of new introductions of alien species per year in Belgium.

# Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

Load libraries:

```{r load_libraries}
library(tidyverse) # To do datascience
library(tidylog) # To provide feedback on dplyr functions
library(rlang) # To use Standard Evaluation in dplyr (!!sym())
library(here) # To find files
library(trias) # To use functions developed for TrIAS
library(INBOtheme) # To use INBO theme for graphs
```

# Get data

```{r get_data_input}
data_file <- here::here(
  "data",
  "interim",
  "data_input_checklist_indicators.tsv"
)
data <- read_tsv(data_file,
  na = "",
  guess_max = 5000
)
```

# Number of new introductions of alien species per year

We use function `indicator_introduction_year` from `trias` library.

## Select data from Belgium

We select data at national level (no regional distributions):

```{r select_data_at_Belgian_level}
data <-
  data %>%
  filter(locationId == "ISO_3166:BE")
```

## Grouping by kingdom

We group data by kingdom:

```{r facet_column_kingdom}
facet_column <- "kingdom"
```

```{r graph_grouping_kingdom, fig.width=10, fig.height=10, results="hide", echo=FALSE}
data %>%
  select(key, facet_column, first_observed) %>%
  distinct() %>%
  indicator_introduction_year(
    start_year_plot = 1920,
    smooth_span = .85,
    x_major_scale_stepsize = 10,
    x_minor_scale_stepsize = 2.5,
    facet_column = facet_column
  )
```

## Grouping by species profile

We group data by habitat:

```{r facet_column_habitat}
facet_column <- "habitat"
```

```{r graph_grouping_habitat, fig.width=10, fig.height=10, results="hide", echo=FALSE}
data %>%
  select(key, facet_column, first_observed) %>%
  distinct() %>%
  indicator_introduction_year(
    start_year_plot = 1920,
    smooth_span = .85,
    x_major_scale_stepsize = 10,
    x_minor_scale_stepsize = 2.5,
    facet_column = facet_column
  )
```

# Checklist indicators for Living Planet Index

In this section we make little changes to the indicator graphs for publication on the Living Planet Index.

## Grouping by kingdom

We group data by kingdom:

```{r facet_column_lpi_kingdom_lpi}
facet_column <- "kingdom"
```

We don't take into account Fungi and Chromista due to a lack of comprehensive data on those kingdoms:

```{r graph_grouping_kingdom_LPI, fig.width=10, fig.height=10, results="hide", echo=FALSE}
data %>%
  select(key, facet_column, first_observed) %>%
  distinct() %>%
  filter(!(!!sym(facet_column)) %in% c("Chromista", "Fungi")) %>%
  indicator_introduction_year(
    start_year_plot = 1920,
    smooth_span = .85,
    x_major_scale_stepsize = 10,
    x_minor_scale_stepsize = 2.5,
    facet_column = facet_column
  )
```

## Grouping by species profile

```{r facet_column_lpi_habitat}
facet_column <- "habitat"
```

We don't take into account combined habitats `freshwater|terrestrial` and `marine|freshwater` due to a lack of comprehensive data on those habitats:

```{r graph_grouping_habitat_LPI, fig.width=10, fig.height=10, results="hide", echo=FALSE}
data %>%
  select(key, facet_column, first_observed) %>%
  distinct() %>%
  filter(!(!!sym(facet_column)) %in% c(
    "freshwater|terrestrial",
    "marine|freshwater",
    NA_character_
  )) %>%
  indicator_introduction_year(
    start_year_plot = 1920,
    smooth_span = .85,
    x_major_scale_stepsize = 10,
    x_minor_scale_stepsize = 2.5,
    facet_column = facet_column
  )
```
