==> src/04_indicator_pathways.Rmd <==
---
title: "Indicator: Pathways associated with alien species introductions"
author:
- Damiano Oldoni
- Stijn Van Hoey
- Tim Adriaens
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
---

This document describes how to build indicators based on checklist data. In particular, this document focuses on pathways of introduction.

# Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      warning = FALSE,
                      message = FALSE,
                      results = FALSE,
                      hide = TRUE
)
```

Load libraries:

```{r load_libraries}
library(tidyverse) # To do datascience
library(tidylog) # To provide feedback on dplyr functions
library(here) # To find files
library(trias) # To use functions developed for TrIAS
library(kableExtra) # To customize tables
library(INBOtheme)  # INBO graph style
```

# Get data

```{r read_data_for_building_indicators}
data_file <- here::here(
  "data",
  "interim",
  "data_input_checklist_indicators.tsv"
)
data <- read_tsv(data_file,
  na = "",
  guess_max = 5000
)
```

# Pathway indicators

## Find categories

We divide the species in categories. Each kingdom, except _Animalia_, is a category. If animal species are present, they are separated in two categories: vertebrates (phylum `Chordata`) and invertebrates (phylum other than `Chordata`, defined as `Not Chordata`).

Categories present in our checklist data:

```{r find_categories_in_data, echo=FALSE, results=TRUE}
categories <- data %>%
  distinct(kingdom) %>%
  rename(category = kingdom)
if ("Animalia" %in% categories$category) {
  if ("Chordata" %in% data$phylum) {
    categories <- categories %>%
      bind_rows(data.frame(
        category = "Chordata",
        stringsAsFactors = FALSE
      ))
  }
  if (data %>%
    filter(kingdom == "Animalia") %>%
    filter(!phylum == "Chordata") %>%
    distinct(phylum) %>% nrow() > 0) {
    categories <- categories %>%
      bind_rows(data.frame(
        category = "Not Chordata",
        stringsAsFactors = FALSE
      ))
  }
  if (any(c("Chordata", "Not Chordata") %in% categories$category)) {
    categories <- categories %>%
      filter(category != "Animalia")
  }
}
categories
```

## Pathway analysis level 1

Total number of taxa introduced for each pathway level 1:

```{r pathway_level1_anaysis}
visualize_pathways_level1(df = data)
```

Per category:

```{r pathway_level1_anaysis_per_kingdom}
map(categories$category, function(x) {
  visualize_pathways_level1(df = data,
                            category = x,
                            title = x)}
)
```

## Pathways analysis level 2

Total number of taxa introduced for pathways level 2:

```{r pathway_level2_anaysis}
pathways_level1 <- 
  data %>%
  distinct(pathway_level1) %>%
  filter(pathway_level1 != "unknown") %>%
  pull()
map(pathways_level1,
    function(x) {
      visualize_pathways_level2(df = data,
                                chosen_pathway_level1 = x,
                                title = x)
    }
)
    
```

Per category:

```{r pathway_level2_anaysis_per_kingdom}
categories_path_lev1 <- expand_grid(categories, pathways_level1)
map2(categories_path_lev1$category,
     categories_path_lev1$pathways_level1,
     function(x, y) {
       visualize_pathways_level2(df = data,
                                 chosen_pathway_level1 = y,
                                 category = x,
                                 title = paste(x, y, collapse = ": "))}
)
```

## Pathways level 1 evolution

We show the evolution in time of the patwhays level 1. 

First, considering all taxa:

```{r visualize_pathways_year_level1_all}
visualize_pathways_year_level1(data)
```

and now splitting them by category:

```{r visualize_pathways_year_level1_category}
map(categories$category, function(x) {
  visualize_pathways_year_level1(df = data,
                                 category = x,
                                 title = x)}
)
```


## Pathways analysis: tables

We also summarize pathway information in one table per category. Number of species for each pathway level 2 and some species as example are provided.

Number of species returned in table as examples of the group:

```{r set_n_species}
n_species <- 5
```

We limit to pathways of taxa introduced from 2000.

```{r year_of_introduction}
start_year <- 2000
```

Create tables:

```{r make_tables_all_data, results="hide"}
tables <- map(categories$category, ~
get_table_pathways(data,
  category = .,
  from = start_year,
  n_species = n_species
))
names(tables) <- categories$category
```

### Pathways per category

Intentionality of introduction is mapped by cell color background: white for intentional introduction (pathway value: `release`), light orange for unintentional introduction.

#### Plantae

Pathways associated to plants (category `Plantae`):

```{r view_tables, echo = FALSE, results = TRUE}
if ("Plantae" %in% categories$category) {
  kable(tables$Plantae) %>%
    row_spec(which(tables$Plantae$pathway_level1 == "release"),
      background = "white"
    ) %>%
    row_spec(which(tables$Plantae$pathway_level1 != "release"),
      background = "#faf0e6"
    ) %>%
    kable_styling() %>%
    scroll_box(width = "100%", height = "10cm")
}
```

#### Vertebrates

Pathways associated to vertebrates (category `Chordata`):

```{r view_table_Chordata, echo = FALSE, results = TRUE}
if ("Chordata" %in% categories$category) {
  kable(tables$Chordata) %>%
    row_spec(which(tables$Chordata$pathway_level1 == "release"),
      background = "white"
    ) %>%
    row_spec(which(tables$Chordata$pathway_level1 != "release"),
      background = "#faf0e6"
    ) %>%
    kable_styling() %>%
    scroll_box(width = "100%", height = "10cm")
}
```

#### Invertebrates

Pathways associated to invertebrates (category `Not Chordata`):

```{r view_table_NotChordata, echo = FALSE, results = TRUE}
if ("Not Chordata" %in% categories$category) {
  kable(tables$`Not Chordata`) %>%
    row_spec(which(tables$`Not Chordata`$pathway_level1 == "release"),
      background = "white"
    ) %>%
    row_spec(which(tables$`Not Chordata`$pathway_level1 != "release"),
      background = "#faf0e6"
    ) %>%
    kable_styling() %>%
    scroll_box(width = "100%", height = "10cm")
}
```

#### Fungi

Pathways associated to fungi (category `Fungi`):

```{r view_table_Fungi, echo = FALSE, results = TRUE}
if ("Fungi" %in% categories$category) {
  kable(tables$Fungi) %>%
    row_spec(which(tables$Fungi$pathway_level1 == "release"),
      background = "white"
    ) %>%
    row_spec(which(tables$Fungi$pathway_level1 != "release"),
      background = "#faf0e6"
    ) %>%
    kable_styling() %>%
    scroll_box(width = "100%", height = "10cm")
}
```

Note: add sections for other kingdoms if needed in the future.

## Table of pathways related to EU policy relevant species

It is important to focus our attention also to a intercategory group of species which is of particular policy importance: the *List of Invasive Alien Species of Union concern* (see more info about [here](http://ec.europa.eu/environment/nature/invasivealien/list/index_en.htm)).

We import this list of species:

```{r}
eu_concern_file <- here::here("data", "input", "eu_concern_species.tsv")
eu_concern_species <- read_tsv(eu_concern_file)
```

Check which species from this subgroup are not present in `data`:

```{r check_presence_eu_species}
eu_concern_species %>%
  filter(!backbone_taxonKey %in% data$nubKey)
```

Categories present in this subgroup:

```{r categories_eu_concern, echo = FALSE, results = TRUE}
categories_eu_concern <- data %>%
  filter(nubKey %in% eu_concern_species$backbone_taxonKey) %>%
  distinct(kingdom) %>%
  rename(category = kingdom)
if ("Animalia" %in% categories_eu_concern$category) {
  if ("Chordata" %in% data$phylum) {
    categories_eu_concern <- categories_eu_concern %>%
      bind_rows(data.frame(
        category = "Chordata",
        stringsAsFactors = FALSE
      ))
  }
  if (data %>%
    filter(kingdom == "Animalia") %>%
    filter(!phylum == "Chordata") %>%
    distinct(phylum) %>% nrow() > 0) {
    categories_eu_concern <- categories_eu_concern %>%
      bind_rows(data.frame(
        category = "Not Chordata",
        stringsAsFactors = FALSE
      ))
  }
  if (any(c("Chordata", "Not Chordata") %in% categories_eu_concern$category)) {
    categories_eu_concern <- categories_eu_concern %>%
      filter(category != "Animalia")
  }
}
categories_eu_concern
```

As done before, we limit to pathways of taxa introduced from 2000.

Create tables:

```{r make_tables_eu_concern, results = "hide"}
tables_eu_concern <- map(categories_eu_concern$category, ~
get_table_pathways(data %>%
  filter(nubKey %in% eu_concern_species$backbone_taxonKey),
category = .,
from = start_year,
n_species = n_species
))
names(tables_eu_concern) <- categories_eu_concern$category
```

### Plantae

Pathways associated to plants (category `Plantae`):

```{r eu_concern_plantae, echo = FALSE, results = TRUE}
if ("Plantae" %in% categories_eu_concern$category) {
  kable(tables_eu_concern$Plantae) %>%
    row_spec(which(tables_eu_concern$Plantae$pathway_level1 == "release"),
      background = "white"
    ) %>%
    row_spec(which(tables_eu_concern$Plantae$pathway_level1 != "release"),
      background = "#faf0e6"
    ) %>%
    kable_styling() %>%
    scroll_box(width = "100%", height = "10cm")
}
```

### Chordata

Pathways associated to vertebrates (category `Chordata`):

```{r eu_concern_chordata, echo = FALSE, results = TRUE}
if ("Chordata" %in% categories_eu_concern$category) {
  kable(tables_eu_concern$Chordata) %>%
    row_spec(which(tables_eu_concern$Plantae$pathway_level1 == "release"),
      background = "white"
    ) %>%
    row_spec(which(tables_eu_concern$Chordata$pathway_level1 != "release"),
      background = "#faf0e6"
    ) %>%
    kable_styling() %>%
    scroll_box(width = "100%", height = "10cm")
}
```

### Not Chordata

Pathways associated to invertebrates (category `Not Chordata`):

```{r eu_concern_not_chordata, echo = FALSE, results = TRUE}
if ("Not Chordata" %in% categories_eu_concern$category) {
  kable(tables_eu_concern$`Not Chordata`) %>%
    row_spec(which(tables_eu_concern$`Not Chordata`$pathway_level1 == "release"),
      background = "white"
    ) %>%
    row_spec(which(tables_eu_concern$`Not Chordata`$pathway_level1 != "release"),
      background = "#faf0e6"
    ) %>%
    kable_styling() %>%
    scroll_box(width = "100%", height = "10cm")
}
```
