==> src/06_occurrence_indicators_appearing_taxa.Rmd <==
---
title: 'Detect appearing and reappearing species'
author:
- Damiano Oldoni
- Toon Van Daele
- Tim Adriaens
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
---

In this document we analyze the Belgian datacube to detect possible (re)appearing taxa in the last year. Such detection is based on some basic rules instead of using GAM (Generalized Additive Models) as the number of GBIF observations decreases sensibly during the current year as effect of a puslishing delay. We also take into account the appearing/reappearing in protected areas as additional warning.

# Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

Load libraries:

```{r load_libraries, message = FALSE}
library(tidyverse) # To do data science
library(tidylog) # To provide feedback on dplyr functions
library(here) # To find files
library(rgbif) # To get information from GBIF
```

# Get data

We read the output of preprocessing pipeline.

Time series:

```{r read_preprocessed_data, message = FALSE}
df_ts <- read_tsv(
  here::here("data", "interim", "df_timeseries.tsv"),
  na = ""
)
```

Columns `pa_obs` and `pa_native_obs` indicate the presence (1) or absence (0) of the specific taxon and native species at class level.

Preview:

```{r preview_ts}
# Get a taxon
taxon <-
  df_ts$taxonKey[10]

# Preview
df_ts %>%
  filter(
    taxonKey == taxon,
    year %in% c(2016, 2017),
    eea_cell_code == "1kmE3924N3102"
  )
```

Taxonomic information:

```{r spec_names}
spec_names <- read_tsv(
  file = here::here(
    "data",
    "interim",
    "timeseries_taxonomic_info.tsv"
  )
)
```

# Detect appearing - reappearing species

Years to evaluate:

```{r eval_years}
present_year <- lubridate::year(Sys.Date())
years <- c(present_year) - 1 # The last full year
years
```

We define the minimum latency (in years) before we speak about reappearance. For example, a latency time window of 3 years means that we consider a taxon reappearing in 2019 if observations occur in 2019 and 2015 or before:

```{r latency_years}
years_latency <- 3
```

## Appearing - reappearing species in Belgium

Lump geographic information:

```{r df_ts_compact}
df_ts_compact <-
  df_ts %>%
  group_by(taxonKey, year, classKey) %>%
  summarise(
    obs = sum(obs),
    cobs = sum(cobs),
    ncells = sum(pa_obs),
    c_ncells = sum(pa_cobs)
  ) %>%
  ungroup()
```

We define the function `is_appearing` for appearing/reappearing detection (this function could be moved to TrIAS package `trias` later):

```{r decision_rules_function}
#' @param df data frame with time series
#' @param eval_years vector with years used for evaluation. Warning is returned
#' if not a sequence. Typically present year and year before.
#' @param latency years of latency (no observations). Strictly positive integer.
#' @param df_reference data frame with time series used to assess reappearing. If NULL (default) df is used.
#' @return data.frame with appearing or reappearing taxa
is_appearing <- function(df,
                         eval_years,
                         latency = 2,
                         df_reference = NULL,
                         data_col = "ncells",
                         id_col = "taxonKey",
                         year_col = "year") {
  df <-
    df %>%
    select(id_col, year_col, data_col) %>%
    rename(
      "taxonKey" = id_col,
      "year" = year_col,
      "occ" = data_col
    ) %>%
    filter(occ > 0) %>%
    group_by(taxonKey)

  if (!is.null(df_reference)) {
    df_reference <-
      df_reference %>%
      select(id_col, year_col, data_col) %>%
      rename(
        "taxonKey" = id_col,
        "year" = year_col,
        "occ" = data_col
      ) %>%
      filter(occ > 0) %>%
      group_by(taxonKey)
  } else {
    df_reference <- df
  }

  # Taxa present in evaluated years

  taxa_eval_years_df <-
    df %>%
    filter(year %in% eval_years) %>%
    filter(year == min(year))

  taxa_eval_years <-
    taxa_eval_years_df %>%
    distinct(taxonKey) %>%
    pull()

  # appearing taxa
  appearing_taxa <-
    df %>%
    filter(taxonKey %in% taxa_eval_years) %>%
    filter(min(year) %in% eval_years) %>%
    filter(year == min(year)) %>%
    ungroup()

  appearing_taxa <-
    appearing_taxa %>%
    rename(!!sym(data_col) := "occ")

  # reappearing taxa
  reappearing_taxa <-
    df_reference %>%
    filter(taxonKey %in% taxa_eval_years) %>%
    filter(year < min(eval_years)) %>%
    filter(year == max(year)) %>%
    ungroup() %>%
    rename(last_year_before_reappearing = year) %>%
    left_join(taxa_eval_years_df, by = "taxonKey") %>%
    group_by(taxonKey) %>%
    summarize(n_latent_years = year - last_year_before_reappearing) %>%
    filter(n_latent_years > latency) %>%
    left_join(taxa_eval_years_df %>% select(taxonKey, occ, year),
      by = "taxonKey"
    )

  reappearing_taxa <-
    reappearing_taxa %>%
    rename(!!sym(data_col) := "occ")

  return(list(
    appearing_taxa = appearing_taxa,
    reappearing_taxa = reappearing_taxa
  ))
}
```

Apply function defined above:

```{r appearing_reappearing_taxa}
appearing_reappearing_taxa <- is_appearing(
  df = df_ts_compact,
  eval_years = years,
  latency = years_latency,
  df_reference = NULL
)
```

Number of appearing taxa:

```{r n_appearing_taxa_BE}
nrow(appearing_reappearing_taxa$appearing_taxa)
```

Number of reappearing taxa:

```{r n_reappearing_taxa_BE}
nrow(appearing_reappearing_taxa$reappearing_taxa)
```

## Appearing - reappearing species in Belgian protected areas

We repeat the workflow defined above to assess which taxa are appearing or reappearing in protected areas.

First we select data referring to protected areas (Natura2000) and then we lump geographical information:

```{r df_ts_compact_protected_areas}
df_ts_compact_prot_areas <-
  df_ts %>%
  filter(natura2000 == TRUE) %>%
  group_by(taxonKey, year) %>%
  summarise(
    obs = sum(obs),
    cobs = sum(cobs),
    ncells = sum(pa_cobs),
    c_ncells = sum(pa_cobs)
  ) %>%
  ungroup()
```

We apply function `is_appearing`. Notice that we use occurrences about all Belgium to assess the reappearance in protected areas:

```{r appearing_reappearing_taxa_prot_areas}
appearing_reappearing_taxa_prot_areas <- is_appearing(
  df = df_ts_compact_prot_areas,
  eval_years = years,
  latency = years_latency,
  df_reference = df_ts_compact
)
```

Number of appearing taxa:

```{r n_appearing_taxa_prot_areas}
nrow(appearing_reappearing_taxa_prot_areas$appearing_taxa)
```

Number of reappearing taxa:

```{r n_appearing_taxa_BE_prot_areas}
nrow(appearing_reappearing_taxa_prot_areas$reappearing_taxa)
```

## Merge

We merge appearing taxa from all Belgium and from protected areas. We rank taxa by area of occupancy (number of cells) in protected areas and in all Belgium. We do the same for reappearing taxa as well:

```{r merge_ranking_by_presence_in protected_area}
app_reapp_global <-
  map2(
    appearing_reappearing_taxa_prot_areas,
    appearing_reappearing_taxa,
    function(x, y) {
      x <-
        x %>%
        rename(ncells_prot_areas = ncells)
      y <-
        y %>%
        rename(ncells_BE = ncells)
      full_join(x, y) %>%
        mutate(
          in_prot_areas = ifelse(!is.na(ncells_prot_areas),
            TRUE,
            FALSE
          ),
          in_BE = ifelse(!is.na(ncells_BE),
            TRUE,
            FALSE
          )
        ) %>%
        arrange(
          desc(ncells_prot_areas),
          desc(ncells_BE)
        )
    }
  )
```

## Add canonical names

Add taxonomic information for better readibility.

```{r add_taxonomic_info}
app_reapp_global <- map(
  app_reapp_global,
  function(x) {
    x %>%
      left_join(spec_names, by = "taxonKey")
  }
)
```

# Table of appearing and reappearing taxa

## Appearing taxa

Reorder columns of data.frame containing appearring taxa:

```{r reorder_cols_appearing}
app_reapp_global$appearing_taxa <-
  app_reapp_global$appearing_taxa %>%
  select(
    taxonKey,
    canonicalName,
    year,
    ncells_prot_areas,
    ncells_BE,
    in_prot_areas,
    in_BE,
    class,
    kingdom,
    classKey,
    kingdomKey
  )
```

Appearing taxa:

```{r add_taxonomic_info_appearing_taxa}
app_reapp_global$appearing_taxa
```

Empty `ncells_BE` and `in_BE` equal to FALSE means that the taxon is appearing in protected areas but not appearing outside, i.e. the taxon is maybe reappearing at Belgian level or occurring with a more stable presence. See next section for more details about. 

## Reappearing taxa

Reorder columns of data.frame containing reappearing taxa:

```{r reorder_cols_reappearing}
app_reapp_global$reappearing_taxa <-
  app_reapp_global$reappearing_taxa %>%
  select(
    taxonKey,
    canonicalName,
    year,
    ncells_prot_areas,
    ncells_BE,
    in_prot_areas,
    in_BE,
    n_latent_years,
    class,
    kingdom,
    classKey,
    kingdomKey
  )
```

Reappearing taxa:

```{r add_taxonomic_info_reappearing_taxa}
app_reapp_global$reappearing_taxa
```

Taxa appearing in protected areas and reappearing outside:

```{r appearing_in_pa_reappearing_outside}
app_reapp_global$appearing_taxa %>%
  filter(in_BE == FALSE) %>%
  filter(taxonKey %in% app_reapp_global$reappearing_taxa$taxonKey) %>%
  select(taxonKey, canonicalName)
```

Taxa appearing in protected areas and with a stable presence outside:

```{r appearing_in_pa_stable_outside}
app_reapp_global$appearing_taxa %>%
  filter(in_BE == FALSE) %>%
  filter(!taxonKey %in% app_reapp_global$reappearing_taxa$taxonKey) %>%
  select(taxonKey, canonicalName, class, kingdom, classKey, kingdomKey)
```

Two examples:

```{r taxa_examples, echo=FALSE}
taxa_examples <- app_reapp_global$appearing_taxa %>%
  filter(in_BE == FALSE) %>%
  filter(!taxonKey %in% app_reapp_global$reappearing_taxa$taxonKey) %>%
  distinct(taxonKey, canonicalName) %>%
  head(n = 2)
taxa_examples %>%
  left_join(df_ts_compact %>%
    select(-starts_with("native")),
  by = "taxonKey"
  ) %>%
  left_join(df_ts_compact_prot_areas %>%
    select(-starts_with("native")) %>%
    rename(
      prot_areas_ncells = ncells,
      prot_areas_obs = obs
    ),
  by = c("taxonKey", "year")
  )
```

# Save data

## Appearing taxa

Save appearing taxa:

```{r save_appearing_taxa}
write_tsv(app_reapp_global$appearing_taxa,
  path = here::here("data", "output", "appearing_taxa.tsv"),
  na = ""
)
```

## Reappearing taxa

Save reappearing taxa:

```{r save_reappearing_taxa}
write_tsv(app_reapp_global$reappearing_taxa,
  path = here::here("data", "output", "reappearing_taxa.tsv"),
  na = ""
)
```
