==> src/08_ranking_emerging_status.Rmd <==
---
title: 'Ranking species by emerging status'
author:
- Damiano Oldoni
- Toon Van Daele
- Tim Adriaens
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
    code_folding: hide
---

This document follows the [modelling pipeline](https://trias-project.github.io/indicators/07_occurrence_indicators_modelling.html) to produce a list of introduced species ranked by emerging status.

# Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      message = FALSE,
                      warning = FALSE)
```

Load libraries:

```{r load_libraries}
library(tidyverse) # To do data science
library(tidylog) # To provide feedback on dplyr functions
library(here) # To find files
```

# Get data

## GAM outputs

We read the emerging status based on GAM, saved as one of the outputs of [modelling pipeline](https://trias-project.github.io/indicators/07_occurrence_indicators_modelling.html):

```{r read_em_summary}
method_gam <- "correct_baseline"
em_gam_occs_be <- read_tsv(
  here::here(
    "data",
    "output",
    "GAM_outputs",
    paste0(
      "output_GAM_observations_belgium_",
      method_gam,
      ".tsv"
    )
  )
) %>%
  mutate(indicator = "observations_BE")

em_gam_occs_pa <- read_tsv(
  here::here(
    "data",
    "output",
    "GAM_outputs",
    paste0(
      "output_GAM_observations_pa_",
      method_gam,
      ".tsv"
    )
  )
) %>%
  mutate(indicator = "observations_pa")

em_gam_occupancy_be <- read_tsv(
  here::here(
    "data",
    "output",
    "GAM_outputs",
    paste0(
      "output_GAM_occupancy_belgium_",
      method_gam,
      ".tsv"
    )
  )
) %>%
  mutate(indicator = "occupancy_BE")

em_gam_occupancy_pa <- read_tsv(
  here::here(
    "data",
    "output",
    "GAM_outputs",
    paste0(
      "output_GAM_occupancy_pa_",
      method_gam,
      ".tsv"
    )
  )
) %>%
  mutate(indicator = "occupancy_pa")

em_gam_list <- list(
  gam_observations_be = em_gam_occs_be,
  gam_observations_pa = em_gam_occs_pa,
  gam_occupancy_be = em_gam_occupancy_be,
  gam_occupancy_pa = em_gam_occupancy_pa
) %>%
  map(~ mutate(
    .data = .,
    model = "GAM"
  ))
```

Preview GAM ouput for modelling number of observations in Belgium:

```{r preview_gam_occs_BE}
em_gam_list$gam_observations_be %>% head(10)
```

Preview GAM ouput for modelling occupancy (number of 1x1km cells) in Belgium:

```{r preview_gam_occupancy_BE}
em_gam_list$gam_occupancy_be %>% head(10)
```

The emerging status (column `em_status`) score is encoded as follows:

- 3: emerging
- 2: potentially emerging
- 1: unclear
- 0: not emerging

As you can see in preview above, GAM could not be applied to all taxa (`em_status` = `NA`). Number of taxon/year combinations without emerging status:

```{r get_number_nas}
sum(map_int(
  em_gam_list,
  function(df) {
    df %>%
      filter(is.na(em_status)) %>%
      nrow()
  }
))
```

In these cases we will use the emerging status as assessed by applying decision rules.

## Decision rules outputs

We read the emerging status based on decision rules, saved as one of the outputs of [modelling pipeline](https://trias-project.github.io/indicators/07_occurrence_indicators_modelling.html):

```{r read_dr_results}
em_decision_rules_occs_BE <- read_tsv(
  here::here(
    "data",
    "output",
    "decision_rules_outputs",
    "output_decision_rules_occs_belgium.tsv"
  ),
  na = ""
) %>%
  mutate(indicator = "observations_BE")

em_decision_rules_occs_pa <- read_tsv(
  here::here(
    "data",
    "output",
    "decision_rules_outputs",
    "output_decision_rules_occs_pa.tsv"
  ),
  na = ""
) %>%
  mutate(indicator = "observations_pa")

em_decision_rules_occupancy_BE <- read_tsv(
  here::here(
    "data",
    "output",
    "decision_rules_outputs",
    "output_decision_rules_occupancy_belgium.tsv"
  ),
  na = ""
) %>%
  mutate(indicator = "occupancy_BE")

em_decision_rules_occupancy_pa <- read_tsv(
  here::here(
    "data",
    "output",
    "decision_rules_outputs",
    "output_decision_rules_occupancy_pa.tsv"
  ),
  na = ""
) %>%
  mutate(indicator = "occupancy_pa")

em_decision_rules_list <- list(
  decision_rules_observations_be = em_decision_rules_occs_BE,
  decision_rules_observations_pa = em_decision_rules_occs_pa,
  decision_rules_occupancy_be = em_decision_rules_occupancy_BE,
  decision_rules_occupancy_pa = em_decision_rules_occupancy_pa
) %>%
  map(~ mutate(
    .data = .,
    model = "decision_rules"
  ))
```

Previewing output of decision rules for occupancy in protected areas:

```{r preview_dr_list}
em_decision_rules_list$decision_rules_occupancy_pa %>% head(10)
```

## Scientific names

Read list of scientific names:

```{r read_taxa_names}
taxa_names <- read_tsv(
  here::here(
    "data",
    "interim",
    "timeseries_taxonomic_info.tsv"
  ),
  na = ""
)
```

These names will be added to ranking for better readibility.

# Preprocessing

Merge results from GAM and decision rules:

```{r concatenate_gam_decision_rules}
em_list <- append(em_gam_list, em_decision_rules_list)
```

Spread columns for ranking:

```{r spread_em_dfs}
# make a data.frame binding all dfs by row
em_df <-
  bind_rows(em_list) %>%
  select(-c(starts_with("growth"), starts_with("dr"), method))

# spread by model
em_df <-
  em_df %>%
  pivot_wider(
    names_from = model,
    id_cols = c(taxonKey, year, indicator),
    values_from = em_status
  )

# assign unique emerging score (if GAM is NA, choose decision rules)
em_df <-
  em_df %>%
  mutate(em = ifelse(is.na(GAM), decision_rules, GAM))

# spread by years
em_df <-
  em_df %>%
  pivot_wider(
    names_from = year,
    values_from = em,
    id_cols = c(taxonKey, indicator),
    names_prefix = paste0("year_")
  )

# spread by indicator
em_df <-
  em_df %>%
  pivot_wider(
    names_from = indicator,
    id_cols = taxonKey,
    values_from = starts_with("year"),
    names_prefix = "em_status_"
  )
```

We also add scientific names:

```{r add_sc_names}
em_df <-
  em_df %>%
  left_join(taxa_names, by = "taxonKey")
```

To rank taxa equally placed, we use the average of the minimal guaranteed growth of number of observations in all Belgium over all the evaluation years. A preview of these values follows:

```{r add_mean_growth_occs_BE}
# add mean growth to em_df
em_df <-
  em_df %>%
  left_join(em_gam_list$gam_observations_be %>%
    group_by(taxonKey) %>%
    summarize(mean_growth = mean(growth, na.rm = TRUE)))
# show preview
em_df %>%
  filter(!is.na(mean_growth)) %>%
  head(n = 10) %>%
  select(taxonKey, canonicalName, mean_growth)
```

# Ranking by emerging status

In this section we create the ranking of introduced species based on the emerging status.

## Hierarchical ranking

The ranking is based on the highest emerging status. The following priority rules are applied, in order of importance:

1. the more recent, the higher priority is
2. emerging statuses in protected areas are more important than the ones defined over entire Belgium
3. emerging statuses from occupancy indicators have priority above the ones from occurrence indicator
4. the higher average minimal guaranteed growth (#occs/year), the higher priority is

```{r rank_basic}
ranking <-
  em_df %>%
  arrange(
    desc(year_2018_em_status_occupancy_pa),
    desc(year_2018_em_status_observations_pa),
    desc(year_2018_em_status_occupancy_BE),
    desc(year_2018_em_status_observations_BE),
    desc(year_2017_em_status_occupancy_pa),
    desc(year_2017_em_status_observations_pa),
    desc(year_2017_em_status_occupancy_BE),
    desc(year_2017_em_status_observations_BE),
    desc(year_2016_em_status_occupancy_pa),
    desc(year_2016_em_status_observations_pa),
    desc(year_2016_em_status_occupancy_BE),
    desc(year_2016_em_status_observations_BE),
    desc(mean_growth)
  ) %>%
  select(
    taxonKey, canonicalName, kingdom, class,
    year_2018_em_status_occupancy_pa,
    year_2018_em_status_observations_pa,
    year_2018_em_status_occupancy_BE,
    year_2018_em_status_observations_BE,
    year_2017_em_status_occupancy_pa,
    year_2017_em_status_observations_pa,
    year_2017_em_status_occupancy_BE,
    year_2017_em_status_observations_BE,
    year_2016_em_status_occupancy_pa,
    year_2016_em_status_observations_pa,
    year_2016_em_status_occupancy_BE,
    year_2016_em_status_observations_BE,
    mean_growth,
    everything()
  )
ranking
```

## Point strategy

We can also rank using a point strategy reflecting the same ranking strategy as before. We apply gain factors to all emerging status scores using the number of observations in 2016 in all Belgium as reference (gain factor  = 1). The following gain factors are applied:

| indicator | year | region | gain factor |
| --- | --- | --- | --- | 
occupancy | 2018 | protected areas | 3
observations | 2018 | protected areas | 2.5
occupancy | 2018 | Belgium | 2.5
observations | 2018 | Belgium | 2
occupancy | 2017 | protected areas | 2.5
observations | 2017 | protected areas | 2
occupancy | 2017 | Belgium | 2
observations | 2017 | Belgium | 1.5
occupancy | 2016 | protected areas | 2
observations | 2016 | protected areas | 1.5
occupancy | 2016 | Belgium | 1.5
observations | 2016 | Belgium | 1

```{r rank_by_em_pts}
ranking_pts <-
  em_df %>%
  mutate(em_pts = year_2018_em_status_occupancy_pa * 3 +
    year_2018_em_status_observations_pa * 2.5 +
    year_2018_em_status_occupancy_BE * 2.5 +
    year_2018_em_status_observations_BE * 2 +
    year_2017_em_status_occupancy_pa * 2.5 +
    year_2017_em_status_observations_pa * 2 +
    year_2017_em_status_occupancy_BE * 2 +
    year_2017_em_status_observations_BE * 1.5 +
    year_2016_em_status_occupancy_pa * 2 +
    year_2016_em_status_observations_pa * 1.5 +
    year_2016_em_status_occupancy_BE * 1.5 +
    year_2016_em_status_observations_BE) %>%
  arrange(
    desc(em_pts),
    desc(mean_growth)
  ) %>%
  select(
    taxonKey, canonicalName, kingdom, class,
    em_pts, mean_growth,
    year_2018_em_status_occupancy_pa,
    year_2018_em_status_observations_pa,
    year_2018_em_status_occupancy_BE,
    year_2018_em_status_observations_BE,
    year_2017_em_status_occupancy_pa,
    year_2017_em_status_observations_pa,
    year_2017_em_status_occupancy_BE,
    year_2017_em_status_observations_BE,
    year_2016_em_status_occupancy_pa,
    year_2016_em_status_observations_pa,
    year_2016_em_status_occupancy_BE,
    year_2016_em_status_observations_BE,
    everything()
  )
ranking_pts
```

# Save results

We save both rankings.

Ranking based on hierarchical strategy is saved in `data/output/ranking_emerging_status_hierarchical_strategy.tsv`:

```{r save_ranking_basic}
write_tsv(ranking,
  here::here("data", "output", "ranking_emerging_status_hierarchical_strategy.tsv"),
  na = ""
)
```

Ranking based on points strategy is saved in `data/output/ranking_emerging_status_points_strategy.tsv`:

```{r save_ranking_pts}
write_tsv(ranking_pts,
  here::here("data", "output", "ranking_emerging_status_points_strategy.tsv"),
  na = ""
)
```
