==> src/09_define_overlay_grid_belgium_with_protected_areas.Rmd <==
---
title: "Get cells of Belgian EEA reference grid containing protected areas"
author: 
  - Damiano Oldoni
  - Tim Adriaens
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
library(knitr)
opts_chunk$set(echo = TRUE,
               warning = FALSE,
               message = FALSE,
               results = TRUE)
```

# Goal

Given the EEA reference grid of Belgium at 1km resolution and the Belgian Natura2000 protected areas, this document assesses the cells covering the protected areas.

# Setup 

```{r load_libraries, message=FALSE, warning=FALSE}
library(tidyverse)  # To do datascience
library(tidylog) # To provide feedback on dplyr functions
library(here) # To find files
library(sf) # To work with geospatial data
library(BelgiumMaps.StatBel)  # To load Belgian administrative boundaries
library(INBOtheme)  # To apply INBO style to plots
library(leaflet)  # To make interactive maps
```

# Read input data

## Import EEA reference grid

We import UTM grid data of Belgium at 1 by 1 km resolution. All grids have CRS (Coordinate Reference System) equal to _Belge Lambert 1972_:
  
```{r import_grids_of_BE_from_EEA}
be_grid <- st_read(here::here("data", "external", "utm1_bel"))
be_grid_crs <- st_crs(be_grid)
be_grid_crs
```

## Read Belgian protected areas Natura2000

GIS data of protected areas as defined in Natura2000 can be found on related webpage of [European Environment Agency](https://www.eea.europa.eu/data-and-maps/data/natura-11#tab-gis-data). We downloaded the GIS data as shapefile in folder `data/external/Natura2000_end2019_Shapefile`:

```{r download_data}
path_natura2000 <- here::here(
  "data",
  "external",
  "Natura2000_end2019_Shapefile"
)
if (!file.exists(path_natura2000)) {
  temp <- tempfile()
  download.file(
    "https://cmshare.eea.europa.eu/s/n5L8Lrs9aYD775S/download",
    temp
  )
  unzip(temp, exdir = path_natura2000)
  unlink(temp)
}
```

We read the shapefile:

```{r protected_areas}
protected_areas <- st_read(path_natura2000)
protected_areas_crs <- st_crs(protected_areas)
```

Check whether grid and protected areas have same CRS:

```{r crs_comparison}
be_grid_crs == protected_areas_crs
```

### Select Belgian protected areas

We are interested in protected areas in Belgium:

```{r select_by_country_Belgium}
protected_areas <-
  protected_areas %>%
  filter(MS == "BE")
```

Summary by site type:

```{r n_areas_per_siteype}
protected_areas %>%
  as_tibble() %>%
  group_by(SITETYPE) %>%
  count()
```

## Read Belgian regions

The Belgian regions are available by package `BelgiumMaps.StatBel`:

```{r be_regions}
data(BE_ADMIN_REGION)
BE_ADMIN_REGION <- st_as_sf(BE_ADMIN_REGION)
```

## Add regional information to Belgian proteted areas

In this section we add three columns to know whether a protected area belong to one or more Belgian regions.

### Flemish protected areas

Add column `flanders` to indicate whether the protected area is a Flemish protected areas (`TRUE`/`FALSE`):

```{r flemis_areas_sitecode}
protected_areas_flanders <- tibble(
  SITECODE = c(
    "BE2200043",
    "BE2200032",
    "BE2100017",
    "BE2300006",
    "BE2500121",
    "BE2100024",
    "BE2400008",
    "BE2200030",
    "BE2200037",
    "BE2200041",
    "BE2200042",
    "BE2200038",
    "BE2101437",
    "BE2219312",
    "BE2500831",
    "BE2100323",
    "BE2100045",
    "BE2100020",
    "BE2200028",
    "BE2400009",
    "BE2200033",
    "BE2300044",
    "BE2400012",
    "BE2200626",
    "BE2200727",
    "BE2200029",
    "BE2400010",
    "BE2223316",
    "BE2301235",
    "BE2301336",
    "BE2300007",
    "BE2422315",
    "BE2524317",
    "BE2100026",
    "BE2100040",
    "BE2200031",
    "BE2500003",
    "BE2200034",
    "BE2100424",
    "BE2400014",
    "BE2101538",
    "BE2200039",
    "BE2500932",
    "BE2500001",
    "BE2218311",
    "BE2101639",
    "BE2217310",
    "BE2300222",
    "BE2400011",
    "BE2500002",
    "BE2220313",
    "BE2200035",
    "BE2100015",
    "BE2300005",
    "BE2301134",
    "BE2100016",
    "BE2100019",
    "BE2200036",
    "BE2500004",
    "BE2200525",
    "BE2501033",
    "BE2221314"
  )) %>%
  mutate(flanders = TRUE,
         wallonia = FALSE,
         brussels = FALSE)
```

### Brussels protected areas

Add column `brussels` to indicate whether the protected area is located in Brussels region (`TRUE`/`FALSE`):

```{r Brussels_areas_sitecode}
protected_areas_brussels <- tibble(
  SITECODE = c("BE1000001",
               "BE1000002",
               "BE1000003")) %>%
  mutate(flanders = FALSE,
         wallonia = FALSE,
         brussels = TRUE)
  
```

### Protected areas under federal administration

Areas not belonging to any region as they are under federal administration:

```{r federal_areas}
protected_areas_federal <- tibble(
  SITECODE = c("BEMNZ0001",
               "BEMNZ0002",
               "BEMNZ0003",
               "BEMNZ0004",
               "BEMNZ0005")) %>%
  mutate(flanders = FALSE,
         wallonia = FALSE,
         brussels = FALSE)
```

### Wallonia protected areas

All the other protected areas are in Wallonia (`wallonia` = `TRUE`):

```{r Wallonia_areas_sitecode}
protected_areas_wallonia <- tibble(
  SITECODE = as.character(protected_areas$SITECODE[
    !protected_areas$SITECODE %in% c(protected_areas_flanders$SITECODE,
                                     protected_areas_brussels$SITECODE,
                                     protected_areas_federal$SITECODE)
  ])) %>%
  mutate(flanders = FALSE,
         wallonia = TRUE,
         brussels = FALSE)
```

### Add regional information

We can now merge the regional information

```{r protected_areas_regional_info}
protected_areas_regional_info <-
  bind_rows(protected_areas_flanders,
            protected_areas_wallonia,
            protected_areas_brussels,
            protected_areas_federal) %>%
  mutate(SITECODE = as.factor(SITECODE))
```

to add it to `protected_areas`:

```{r add_regional_membership}
protected_areas <- 
  protected_areas %>%
  left_join(protected_areas_regional_info,
            by = "SITECODE")
```

Preview:

```{r preview_regions}
protected_areas %>%
  as_tibble() %>%
  select(SITECODE, SITENAME, SITETYPE, flanders, wallonia, brussels) %>%
  head()
```

### Distribution of Belgian protected areas

Number of protected areas per type:

```{r n_prot_area_per_type}
protected_areas %>%
  st_drop_geometry() %>%
  group_by(SITETYPE) %>%
  count()
```

In Flanders:

```{r n_prot_area_per_type_flanders}
protected_areas %>%
  filter(flanders == TRUE) %>%
  st_drop_geometry() %>%
  group_by(SITETYPE) %>%
  count()
```

```{r leaflet_prot_areas_flanders, cache = TRUE}
# Transform to wgs84
protected_areas_wgs84 <- 
  protected_areas %>%
          st_transform(crs = 4326)
# Flemish protected areas
protected_areas_wgs84_flanders <- 
  protected_areas_wgs84 %>%
  filter(flanders == TRUE)

prot_area_fl_palette <- colorFactor(
  topo.colors(nrow(protected_areas_flanders)),
  protected_areas_flanders$SITECODE)
leaflet() %>%
  addTiles() %>%
  addPolygons(data = protected_areas_wgs84_flanders %>%
                filter(SITETYPE == "A"),
              fillColor = ~prot_area_fl_palette(SITECODE),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~SITENAME,
              popup = ~SITECODE,
              group = "type A") %>%
  addPolygons(data = protected_areas_wgs84_flanders %>%
                filter(SITETYPE == "B"),
              fillColor = ~prot_area_fl_palette(SITECODE),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~SITENAME,
              popup = ~SITECODE,
              group = "type B") %>%
  addLayersControl(
    overlayGroups = c("type A", "type B"),
    options = layersControlOptions(collapsed = FALSE))
```

In Wallonia:

```{r n_prot_area_per_type_wallonia}
protected_areas %>%
  filter(wallonia == TRUE) %>%
  st_drop_geometry() %>%
  group_by(SITETYPE) %>%
  count()
```

```{r leaflet_prot_areas_wa, cache = TRUE}
protected_areas_wgs84_wallonia <- 
  protected_areas_wgs84 %>%
  filter(wallonia == TRUE)

prot_area_wallonia_palette <- colorFactor(
  topo.colors(nrow(protected_areas_wallonia)),
  protected_areas_wallonia$SITECODE)
leaflet() %>%
  addTiles() %>%
  addPolygons(data = protected_areas_wgs84_wallonia %>%
                filter(SITETYPE == "A"),
              fillColor = ~prot_area_wallonia_palette(SITECODE),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~SITENAME,
              popup = ~SITECODE,
              group = "type A") %>%
  addPolygons(data = protected_areas_wgs84_wallonia %>%
                filter(SITETYPE == "B"),
              fillColor = ~prot_area_wallonia_palette(SITECODE),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~SITENAME,
              popup = ~SITECODE,
              group = "type B") %>%
  addPolygons(data = protected_areas_wgs84_wallonia %>%
                filter(SITETYPE == "C"),
              fillColor = ~prot_area_wallonia_palette(SITECODE),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~SITENAME,
              popup = ~SITECODE,
              group = "type C") %>%
  addLayersControl(
    overlayGroups = c("type A", "type B", "type C"),
    options = layersControlOptions(collapsed = FALSE))
```

In Brussels:

```{r n_prot_area_per_type_brussels}
protected_areas %>%
  filter(brussels == TRUE) %>%
  st_drop_geometry() %>%
  group_by(SITETYPE) %>%
  count()
```

```{r leaflet_prot_areas_br, cache = TRUE}
protected_areas_wgs84_brussels <- 
  protected_areas_wgs84 %>%
  filter(brussels == TRUE)

prot_area_brussels_palette <- colorFactor(
  topo.colors(nrow(protected_areas_brussels)),
  protected_areas_brussels$SITECODE)
leaflet() %>%
  addTiles() %>%
  addPolygons(data = protected_areas_wgs84_brussels %>%
                filter(SITETYPE == "B"),
              fillColor = ~prot_area_brussels_palette(SITECODE),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~SITENAME,
              popup = ~SITECODE,
              group = "type B") %>%
  addLayersControl(
    overlayGroups = c("type B"),
    options = layersControlOptions(collapsed = FALSE))
```

Areas under federal administration

```{r n_prot_area_per_type_fed}
protected_areas %>%
  filter(flanders == FALSE,
         wallonia == FALSE,
         brussels == FALSE) %>%
  st_drop_geometry() %>%
  group_by(SITETYPE) %>%
  count()
```

```{r leaflet_prot_areas_fed_admin, hide = TRUE, cache = TRUE}
# Protected areas under federal administration
protected_areas_wgs84_federal <- 
  protected_areas_wgs84 %>%
  filter(flanders == FALSE,
         wallonia == FALSE,
         brussels == FALSE)

prot_area_palette_federal <- colorFactor(
  topo.colors(nrow(protected_areas_federal)),
  protected_areas_federal$SITECODE)
leaflet() %>%
  addTiles() %>%
  addPolygons(data = protected_areas_wgs84_federal %>%
                filter(SITETYPE == "A"),
              fillColor = ~prot_area_palette_federal(SITECODE),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~SITENAME,
              popup = ~SITECODE,
              group = "type A") %>%
  addPolygons(data = protected_areas_wgs84_federal %>%
                filter(SITETYPE == "B"),
              fillColor = ~prot_area_palette_federal(SITECODE),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~SITENAME,
              popup = ~SITECODE,
              group = "type B") %>%
  addLayersControl(
    overlayGroups = c("type A", "type B"),
    options = layersControlOptions(collapsed = FALSE))
```

# Define Bird and Habitat directive areas

## Bird directive areas

Bird directive areas (SPA areas) are defined as the areas with `SITETYPE`: A and C.

```{r bird_directive_areas}
protected_areas_bird <-
  protected_areas %>%
  filter(SITETYPE %in% c("A", "C"))
```

## Habitat directive areas

Habitat directive areas (SAC areas) are defined as the protected areas with `SITETYPE` values B and C.

```{r habitat_directive_areas}
protected_areas_habitat <-
  protected_areas %>%
  filter(SITETYPE %in% c("B", "C"))
```

# Intersect grid of Belgium with protected areas

## Interset grid with protected areas

Intersect cells of the grid and protected areas:

```{r intersect_grid_protected_areas}
grid_intersects_areas <- st_intersects(be_grid, protected_areas,
  sparse = FALSE
)
grid_intersects_areas <- as_tibble(grid_intersects_areas,
  .name_repair = "universal"
)
# Assign column names based on protected areas code
prot_areas_sitecode <- as.character(protected_areas$SITECODE)
names(grid_intersects_areas) <- prot_areas_sitecode
# Add 1x1km grid cellcode as new column
grid_intersects_areas <-
  grid_intersects_areas %>%
  mutate(CELLCODE = as.character(droplevels(be_grid$CELLCODE))) %>%
  select(CELLCODE, everything())
```

## Natura2000 protected areas

Select cells which intersect any of the protected areas:

```{r cells_with_protected_areas}
cells_protected <-
  grid_intersects_areas %>%
  filter_at(vars(starts_with("BE")), any_vars(. == TRUE))
```

## Get grid cells for each protected area

For each protected area we can get the list of intersecting cells:

```{r get_list_cells_per_area}
cells_of_prot_areas <-
  map_dfr(
    prot_areas_sitecode,
    function(pa_cell_code) {
      site_type <-
        protected_areas %>%
        dplyr::filter(SITECODE == pa_cell_code) %>%
        dplyr::pull(SITETYPE)
      cells <-
        cells_protected %>%
        # avoid long logs by using dplyr functions instead of tidylog
        dplyr::select(CELLCODE, !!sym(pa_cell_code)) %>%
        dplyr::filter(!!sym(pa_cell_code) == TRUE) %>%
        dplyr::mutate(
          SITECODE = pa_cell_code,
          SITETYPE = as.character(droplevels(site_type))
        ) %>%
        dplyr::select(SITECODE, SITETYPE, CELLCODE)
    }
  )
```

For example, grid cells intersecting protected area `BE1000003`:

```{r example_BE31001C0}
cells_of_prot_areas %>%
  filter(SITECODE == "BE1000003")
```

Distribution of the number of cells covering the protected areas:

```{r distribution_n_cells_protected_areas}
cells_of_prot_areas %>%
  group_by(SITECODE) %>%
  count() %>%
  arrange(desc(n)) %>%
  ggplot() +
  geom_histogram(aes(x = n), bins = 30) +
  scale_x_log10() +
  xlab("Number of cells (km2)") +
  ylab("Number of protected areas")
```

## Natura2000

We can now define whether a cell of the reference grid interesects a protected area as defined by Natura2000. We add this information in column `natura2000`:

```{r cells_natura2000}
cellcode_protected_natura2000 <-
  cells_protected %>%
  as_tibble() %>%
  pull(CELLCODE)

be_grid_membership_protected_areas <-
  be_grid %>%
  mutate(natura2000 = ifelse(CELLCODE %in% cellcode_protected_natura2000,
    TRUE,
    FALSE
  ))
```

### Bird directive areas

We add also a column, `spa`, assessing whether the cell intersects a SPA area:

```{r be_grid_add_spa}
cells_protected_spa <-
  cells_protected %>%
  select(
    CELLCODE,
    one_of(as.character(protected_areas_bird$SITECODE))
  ) %>%
  filter_at(vars(starts_with("BE")), any_vars(. == TRUE)) %>%
  as_tibble() %>%
  pull(CELLCODE)

be_grid_membership_protected_areas <-
  be_grid_membership_protected_areas %>%
  mutate(spa = ifelse(CELLCODE %in% cells_protected_spa,
    TRUE,
    FALSE
  ))
```

### Habitat directive areas

We add also a column, `habitat`, assessing whether the cell intersects a habitat directive area:

```{r be_grid_add_directive}
cells_protected_hab <-
  cells_protected %>%
  select(
    CELLCODE,
    one_of(as.character(protected_areas_habitat$SITECODE))
  ) %>%
  filter_at(vars(starts_with("BE")), any_vars(. == TRUE)) %>%
  as_tibble() %>%
  pull(CELLCODE)

be_grid_membership_protected_areas <-
  be_grid_membership_protected_areas %>%
  mutate(habitat = ifelse(CELLCODE %in% cells_protected_hab,
    TRUE,
    FALSE
  ))
```

## Summary

How many cells contain/intersect protected areas?

```{r summary_natura2000}
be_grid_membership_protected_areas <-
  be_grid_membership_protected_areas %>%
  as_tibble()
be_grid_membership_protected_areas %>%
  group_by(natura2000) %>%
  count()
```

How many cells contain/intersect SPA areas?

```{r summary_spa_areas}
be_grid_membership_protected_areas %>%
  group_by(spa) %>%
  count()
```

How many cells contain/intersect habitat directive areas?

```{r summary_spa_habitat}
be_grid_membership_protected_areas %>%
  group_by(habitat) %>%
  count()
```

## Save data

We save `be_grid_membership_protected_areas` as tab-separated file (tsv):

```{r save_data}
write_tsv(be_grid_membership_protected_areas,
  path = here::here(
    "data",
    "interim",
    "intersect_EEA_ref_grid_protected_areas.tsv"
  ),
  na = ""
)
```

We also save `cells_of_prot_areas` in the same format:

```{r save_cells_of_prot_areas}
write_tsv(cells_of_prot_areas,
  path = here::here(
    "data",
    "interim",
    "EEA_ref_grid_cells_covering_protected_areas.tsv"
  ),
  na = ""
)
```

We save the Belgian protected areas as a geopackage as it will be used elsewhere:

```{r save_protected_areas}
protected_areas %>%
  st_write(
    here::here("data",
               "output",
               "Belgian_Natura2000_protected_areas.gpkg"),
    delete_dsn =TRUE
)
```

We are also interested in saving basic information about protected areas:

```{r save_metadata_prot_areas}
protected_areas_metadata <-
  protected_areas %>%
  st_drop_geometry() %>%
  select(SITECODE,
         SITENAME,
         SITETYPE,
         flanders,
         wallonia,
         brussels)
protected_areas_metadata
```

We save these metadata:

```{r save_protected_areas_metadata}
protected_areas_metadata %>%
  write_tsv(
    path = here::here(
      "data",
      "interim",
      "protected_areas_metadata.tsv"
    ),
    na = ""
  )
```
