==> src/10_species_observations_occupancy_in_protected_areas.Rmd <==
---
title: "Species observations and occupancy in Belgian protected areas"
author: 
  - Damiano Oldoni
  - Tim Adriaens
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
library(knitr)
opts_chunk$set(echo = TRUE,
               warning = FALSE,
               message = FALSE,
               results = TRUE)
```

# Goal

In this document we calculate the area of occupancy (AOO), the coverage and the number of observations in Belgian natura2000 protected areas. Data are aggregated per year, species and at a resolution of 1 km x 1km as we use the EEA reference grid for Belgium.

# Setup 

```{r load_libraries, message=FALSE, warning=FALSE}
library(tidyverse)  # To do datascience
library(tidylog)  # To provide feedback on dplyr functions
library(here) # To find files
library(inborutils)
library(rgbif)
```

# Read input data

## Occurrence cube at species level

Read the occurrence cubes at species level for European countries as published on Zenodo (https://zenodo.org/record/3637911). It can take a long time:

```{r read_cube_from_zenodo}
if (!"be_species_cube.csv" %in%
  list.files(here::here("data", "interim"))) {
  download_zenodo(
    doi = "10.5281/zenodo.3637911",
    path = here::here("data", "interim")
  )
}
```

We are interested in the Belgian cube (`be_species_cube.csv`) and its metadata (`be_species_info.csv`):

```{r read_be_species_cube}
be_species_cube <- read_csv(
  file = here::here("data", "interim", "be_species_cube.csv"),
  na = ""
)
be_species_info <- read_csv(
  file = here::here("data", "interim", "be_species_info.csv"),
  na = ""
)
```

Preview occurrence cube:

```{r preview_cube}
be_species_cube %>% head(10)
```

Preview taxonomic metadata of occurrence cube:

```{r preview_cube_metadata}
be_species_info %>% head(10)
```

## Occurrence cube of alien taxa

Read the occurrence cubes of alien taxa as published on Zenodo (https://zenodo.org/record/3637911). It is based on GBIF occurrence data of taxa published in the [Global Register of Introduced and Invasive Species - Belgium](https://www.gbif.org/dataset/6d9e952f-948c-4483-9807-575348147c7e), shortly called the [unified checklist](https://github.com/trias-project/unified-checklist). Attention: It can take a long time:

```{r read_cube_from_zenodo_alien}
if (!"be_alientaxa_cube.csv" %in%
  list.files(here::here("data", "interim"))) {
  download_zenodo(
    doi = "10.5281/zenodo.3635510",
    path = here::here("data", "interim")
  )
}
```

We are interested in the alien taxa cube (`be_alientaxa_cube.csv`) and its metadata (`be_alientaxa_info.csv`):

```{r read_be_species_cube_alien}
be_alientaxa_cube <- read_csv(
  file = here::here("data", "interim", "be_alientaxa_cube.csv"),
  na = ""
)
be_alientaxa_info <- read_csv(
  file = here::here("data", "interim", "be_alientaxa_info.csv"),
  na = ""
)
```

Preview occurrence cube:

```{r preview_cube_alien}
be_alientaxa_cube %>% head(10)
```

Preview taxonomic metadata of occurrence cube with alien taxa:

```{r preview_cube_metadata_alien}
be_alientaxa_info %>% head(10)
```

Number of taxa per rank:

```{r rank_alien}
be_alientaxa_info %>%
  distinct(rank, taxonKey) %>%
  group_by(rank) %>%
  count()
```

Infraspecific alien taxa:

```{r species_with_alien_infraspecific_taxa, cache = TRUE}
infraspecific_alien <-
  be_alientaxa_info %>%
  distinct(rank, taxonKey) %>%
  filter(rank %in% c("SUBSPECIES",
                     "VARIETY",
                     "FORM"))
infraspecific_alien <-
  map_dfr(
    infraspecific_alien$taxonKey,
    function(x) {
      name_usage(x, return = "data")
    }
  )
infraspecific_alien <-
  infraspecific_alien %>%
  mutate(
    species = if_else(is.na(species),
      parent,
      species
    ),
    speciesKey = if_else(is.na(speciesKey),
      parentKey,
      speciesKey)) %>%
  select(taxonKey = key,
         kingdom,
         phylum,
         order,
         class,
         family,
         genusKey,
         genus,
         speciesKey,
         species)
infraspecific_alien <-
  infraspecific_alien %>%
  left_join(be_alientaxa_info,
            by = c("taxonKey"))
infraspecific_alien
```

## Assessment alien species with alien infraspecific taxa

The unified checklist contains some infraspecific alien taxa belonging to alien species as well. The assessment whether the species is alien or not is at the moment based on a reference file maintained by TrIAS experts. We read it:

```{r infraspecific_alien_status_species}
infraspecific_alien_status_species <-
  read_tsv(here::here("reference",
                      "species_of_infraspecific_alien_taxa.tsv"),
           na = "")
infraspecific_alien_status_species
```

where `key` is the key of the taxon as published in the unified checklist,  `nubKey` is the key of the corresponding taxon in the GBIF Backbone. Also the columns `speciesKey`, `species`, `scientificName` and  `kingdom` contain information coming from the GBIF Backbone.

## EEA cells containing the Natura2000 protected areas

We read also the file containing the intersection of protected areas and the EEA grid as produced in :

```{r read_cells_of_prot_areas}
cells_of_prot_areas <- read_tsv(
  file = here::here(
    "data",
    "interim",
    "EEA_ref_grid_cells_covering_protected_areas.tsv"
  ),
  na = ""
)
```

Preview:

```{r preview_cells_of_prot_areas}
cells_of_prot_areas %>% head(10)
```

Number of cells per protected area:

```{r n_cells_per_prot_areas}
n_cells_prot_areas <-
  cells_of_prot_areas %>%
  group_by(SITECODE) %>%
  summarize(n_cells = n())
```

Preview:

```{r preview_n_cells_prot_areas}
n_cells_prot_areas %>% head(10)
```

## Protected areas metadata

Metadata about all Belgian Natura2000 protected areas have been already saved in [other pipeline](define_overlay_grid_belgium_with_protected_areas.html). We read them:

```{r protected_areas_metadata}
protected_areas_metadata <- read_tsv(
  here::here(
    "data",
    "interim",
    "protected_areas_metadata.tsv"
  ),
  na = ""
)
```

# Calculate number of observations, AOO and coverage of Belgian protected areas

Get values from occurrence cube related to cells covering the protected areas:

```{r join_cell_prot_areas_with_be_cube}
be_prot_areas_cube <-
  cells_of_prot_areas %>%
  inner_join(be_species_cube,
    by = c("CELLCODE" = "eea_cell_code")
  )
```

Preview:

```{r preview_be_prot_areas_cube}
be_prot_areas_cube %>% head(10)
```

For each taxon, year and protected area, calculate the number of observations, the area of occupancy and the minimum coordinate uncertainty:

```{r occs_prot_areas}
occs_prot_areas <-
  be_prot_areas_cube %>%
  group_by(SITECODE, SITETYPE, year, speciesKey) %>%
  summarize(
    aoo = n(),
    n = sum(n),
    min_coord_uncertainty = min(min_coord_uncertainty)
  ) %>%
  ungroup()
```

Preview with the 10 highest area of occupancies:

```{r preview_occs_prot_areas}
occs_prot_areas %>%
  arrange(desc(aoo)) %>%
  head(10)
```

## Add coverage of protected areas

To add the coverage of each protected area, we divide AOO by the area of each protected area defined as the number of cells totally covering the protected area:

```{r add_coverage}
occs_prot_areas <-
  occs_prot_areas %>%
  left_join(n_cells_prot_areas,
    by = c("SITECODE")
  ) %>%
  mutate(coverage = aoo / n_cells) %>%
  select(-n_cells)
```

The coverage is a number between 0 (species not present) and 1 (species present in the entire protected area). Species/year/area with highest coverage value (1):

```{r preview_coverage}
occs_prot_areas %>%
  arrange(desc(coverage)) %>%
  filter(coverage == max(coverage))
```

## Retrieve taxonomic metadata from GBIF

Not all taxa in Belgian occurrence cube are present in protected areas. We remove them from the taxonomic metadata of the occurrence cube:

```{r reduce_be_species_info}
prot_areas_species_info <-
  be_species_info %>%
  filter(speciesKey %in% occs_prot_areas$speciesKey)
```

We retrieve taxonomic tree from GBIF:

```{r taxonomic_info_species_cube, cache = TRUE}
prot_areas_species_taxon_info <-
  map_dfr(
    prot_areas_species_info$speciesKey,
    function(x) {
      name_usage(key = x, return = "data")
    }
  )
prot_areas_species_taxon_info <-
  prot_areas_species_taxon_info %>%
  select(
    speciesKey,
    kingdom,
    phylum,
    order,
    class,
    family,
    genusKey,
    genus,
    species
  )
prot_areas_species_info <-
  prot_areas_species_info %>%
  left_join(prot_areas_species_taxon_info,
    by = "speciesKey"
  )
```

Preview:

```{r preview_prot_areas_species_taxon_info_after_gbif}
prot_areas_species_taxon_info %>% head(10)
```

## Add alien status

We would also like to indicate whether the species is alien. We add a boolean column indicating whether the species is alien or not based on the unified checklist:

```{r add_is_alien}
# alien species based on alien infraspecific taxa in unified
infraspecific_alien_and_species <- 
  infraspecific_alien_status_species %>%
  filter(species_is_alien == TRUE)

infraspecific_alien_species_unknown <-
  infraspecific_alien_status_species %>%
  filter(is.na(species_is_alien))

occs_prot_areas <-
  occs_prot_areas %>%
  mutate(is_alien = case_when(
    speciesKey %in% be_alientaxa_info$taxonKey |
      speciesKey %in% infraspecific_alien_and_species$speciesKey ~ TRUE,
    speciesKey %in% 
      infraspecific_alien_species_unknown$speciesKey ~ NA,
    TRUE ~ FALSE
  ))
```

As the Belgian list of alien taxa contains infraspecific taxa, we add a remark for the corresponding species:

```{r add_remarks}
species_infraspecific_alien <-
  infraspecific_alien %>%
  select(speciesKey, scientificName, taxonKey) %>%
  filter(speciesKey %in% occs_prot_areas$speciesKey) %>%
  group_by(speciesKey) %>%
  summarize(remarks = paste0(
    "Infraspecific alien taxa present. ",
    paste(taxonKey,
      scientificName,
      sep = ":",
      collapse = ","
    )
  ))
occs_prot_areas <-
  occs_prot_areas %>%
  left_join(species_infraspecific_alien,
    by = "speciesKey"
  )
```

Species with alien infraspecific taxa. Notice they can be considered alien as well:

```{r example_alien_infraspecific}
occs_prot_areas %>%
  filter(!is.na(remarks)) %>%
  distinct(speciesKey, is_alien, remarks)
```

Finally we add a column, `taxonKey` containing the key of all taxa in GBIF Backbone. It is by definition equal to `speciesKey`.

```{r add_species_key}
occs_prot_areas <-
  occs_prot_areas %>%
  mutate(taxonKey = speciesKey)
prot_areas_species_info <-
  prot_areas_species_info %>%
  mutate(taxonKey = speciesKey)
```

## Include infraspecific alien taxa

To better study the spread of alien taxa we also add information about infraspecific alien taxa:

```{r be_prot_areas_cube_infraspecific}
# Select cells with occurrences of infraspecific alien taxa
be_prot_areas_cube_infraspecific <-
  cells_of_prot_areas %>%
  inner_join(
    be_alientaxa_cube %>%
      filter(taxonKey %in% infraspecific_alien$taxonKey),
    by = c("CELLCODE" = "eea_cell_code")
  )
# Calculate n observations, aoo, min_coord_uncertainty
occs_prot_areas_infraspecific <-
  be_prot_areas_cube_infraspecific %>%
  group_by(SITECODE, SITETYPE, year, taxonKey) %>%
  summarize(
    aoo = n(),
    n = sum(n),
    min_coord_uncertainty = min(min_coord_uncertainty)
  ) %>%
  ungroup()
# Add coverage and is_alien
occs_prot_areas_infraspecific <-
  occs_prot_areas_infraspecific %>%
  left_join(n_cells_prot_areas,
    by = c("SITECODE")
  ) %>%
  mutate(coverage = aoo / n_cells,
         is_alien = TRUE) %>%
  select(-n_cells)
# Remove taxa which are not present in protected areas
prot_areas_infraspecific_taxon_info <-
  infraspecific_alien %>%
  filter(taxonKey %in% occs_prot_areas_infraspecific$taxonKey)
# Add speciesKey
occs_prot_areas_infraspecific <-
  occs_prot_areas_infraspecific %>%
  left_join(infraspecific_alien %>%
              select(taxonKey, speciesKey),
            by = "taxonKey")
# Merge data
occs_prot_areas <-
  occs_prot_areas %>%
  bind_rows(occs_prot_areas_infraspecific)
# Merge metadata
prot_areas_taxon_info <-
  prot_areas_species_info %>%
  bind_rows(prot_areas_infraspecific_taxon_info)
```

# Save data and metadata

## Remove erroneous taxa

As discussed in [issue #32](https://github.com/trias-project/unified-checklist/issues/32) we remove the following species:

- [Anser fabalis (Latham, 1787)](https://www.gbif.org/species/2498027)
- [Anser anser (Linnaeus, 1758)](https://www.gbif.org/species/2498036)
- [Branta leucopsis (Bechstein, 1803)](https://www.gbif.org/species/5232464)

```{r taxa_to_remove}
taxa_to_remove_speciesKey <- c(
  # Anser fabalis (Latham, 1787)
  2498027, 
  # Anser anser (Linnaeus, 1758)
  2498036,
  # Branta leucopsis (Bechstein, 1803)
  5232464
)
```

Important: this step is a patch waiting for a new version of the [Global Register of Introduced and Invasive Species - Belgium](https://www.gbif.org/dataset/6d9e952f-948c-4483-9807-575348147c7e).

```{r remove_taxa}
# Remove from data
occs_prot_areas <-
  occs_prot_areas %>%
  filter(!taxonKey %in% taxa_to_remove_speciesKey)
# Remove from metadata
prot_areas_taxon_info <-
  prot_areas_taxon_info %>%
  filter(!taxonKey %in% taxa_to_remove_speciesKey)
```

## Save data about protected area 

We finalize the data.frame containing the information about number of observations, occupancy, coverage and alien status in Belgian protected areas.

Select column of interest and set columns order:

```{r finalize_occs_prot_areas}
occs_prot_areas <-
  occs_prot_areas %>%
  select(
    SITECODE,
    SITETYPE,
    year,
    taxonKey,
    speciesKey,
    n,
    aoo,
    coverage,
    min_coord_uncertainty,
    is_alien,
    remarks
  )
```

Preview:

```{r preview_occs_prot_areas_before_saving}
occs_prot_areas %>% head(10)
```

Save data:

```{r save_occs_prot_areas}
occs_prot_areas %>%
  write_csv(
    path = here::here(
      "data",
      "output",
      "protected_areas_species_occurrence.csv"
    ),
    na = ""
  )
```

## Finalize en save metadata

### Taxonomic metadata

Select taxonomic metadata we are interested to:

```{r selection_taxon_metadata}
taxonomic_metadata <-
  prot_areas_taxon_info %>%
  select(
    taxonKey,
    speciesKey,
    scientificName,
    kingdom,
    phylum,
    order,
    class,
    genus,
    family,
    species,
    rank,
    includes
  )
```

Preview:

```{r preview_taxonomic_metadata_before_saving}
taxonomic_metadata %>% head(10)
```

Save taxonomic metadata:

```{r save_taxonomic_metadata}
taxonomic_metadata %>%
  write_csv(here::here(
    "data",
    "output",
    "protected_areas_species_info.csv"
  ),
  na = ""
  )
```

### Protected areas metadata

We remove metadata of areas without observations, if any:

```{r filter_protected_areas_metadata}
protected_areas_metadata <-
  protected_areas_metadata %>%
  filter(SITECODE %in% occs_prot_areas$SITECODE)
```

Preview:

```{r preview_protected_areas_metadata_before_saving}
protected_areas_metadata %>% head(10)
```

We save the metadata as output:

```{r save_final_protected_areas_metadata}
protected_areas_metadata %>%
  write_csv(
    path = here::here(
      "data",
      "output",
      "protected_areas_metadata.csv"
    ),
    na = ""
  )
```
