==> src/11_status_alien_species_in_protected_areas.Rmd <==
---
title: "Status of alien species in Belgian protected areas"
author: 
  - Damiano Oldoni
  - Tim Adriaens
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
library(knitr)
opts_chunk$set(echo = TRUE,
               warning = FALSE,
               message = FALSE,
               results = TRUE)
```

# Goal

In this document we analyze the status of alien species in the Belgian protected areas.

# Setup 

```{r load_libraries, message=FALSE, warning=FALSE}
library(tidyverse)  # To do datascience
library(lubridate)  # To work with dates
library(tidylog)  # To provide feedback on dplyr functions
library(here) # To find files
library(purrr) # To do loops
library(sf) # To work with geospatial data
library(leaflet) # To make interactive maps
```

# Read input data

Read data related to species distribution in protected areas:

```{r read_occs_prot_areas}
occs_prot_areas <-
  read_csv(here::here(
    "data",
    "output",
    "protected_areas_species_occurrence.csv"
  ), na = ""
)
```

Related taxonomic metadata:

```{r read_taxonomic_metadata}
taxonomic_metadata <-
  read_csv(here::here("data",
                      "output",
                      "protected_areas_species_info.csv"),
  na = "",
  col_types = cols(
    speciesKey = col_double()
  )
)
```

and protected areas metadata:

```{r read_protected_areas_metadata}
protected_areas_metadata <-
  read_csv(
    here::here(
      "data",
      "output",
      "protected_areas_metadata.csv"
    ),
    na = ""
)
```

Read the geospatial package containig the Belgian protected areas:

```{r read_protected_areas}
protected_areas <-
  st_read(here::here("data",
                     "output",
                     "Belgian_Natura2000_protected_areas.gpkg")
)
```

Read the list of invasive alien species of Union concern:

```{r eu_list}
eu_concern_species <- 
  read_tsv(
    here::here(
      "data",
      "input",
      "eu_concern_species.tsv"
    ),
  na = ""
)
```


# Stage of invasion of protected areas

We analyze the stage of invasion of protected areas based on observations from 2000.

```{r begin_year}
begin_year <- 2000
```

## Natura2000 areas

We start by taking into account all type of areas.

### Belgium

Number of alien taxa and total number of observations of alien taxa in each protected area from 2000 up to now:

```{r n_taxa_obs_per_area}
n_taxa_obs_per_area <-
  occs_prot_areas %>%
    left_join(protected_areas_metadata %>% 
                select(-SITETYPE),
              by = "SITECODE") %>%
  filter(is_alien == TRUE) %>%
  filter(year >= begin_year) %>%
  rename(n_obs = n) %>%
  group_by(speciesKey,
           year,
           SITECODE,
           SITENAME,
           SITETYPE,
           flanders,
           wallonia,
           brussels) %>%
  add_tally() %>%
  # avoid to count alien species when infraspecific alien taxa are present
  filter(n == 1 | taxonKey != speciesKey) %>%
  ungroup() %>%
  select(-n) %>%
  group_by(SITECODE,
           SITENAME,
           SITETYPE,
           flanders,
           wallonia,
           brussels) %>%
  summarize(n_taxa =  length(unique(taxonKey)),
            n_obs = sum(n_obs)) %>%
  ungroup() %>%
  arrange(desc(n_taxa), desc(n_obs)) %>%
  select(SITECODE,
         SITENAME,
         SITETYPE,
         n_taxa,
         n_obs,
         everything())
n_taxa_obs_per_area
```

Map with number of alien taxa per protected area:

```{r leaflet_n_taxa_in_prot_aras, echo = FALSE}
# Transform to wgs84
protected_areas_wgs84 <- 
  protected_areas %>%
          st_transform(crs = 4326)

protected_areas_wgs84 <-
  protected_areas_wgs84 %>%
  merge(n_taxa_obs_per_area %>%
              select(SITECODE, n_taxa, n_obs),
        all = TRUE)

# bins <- c(0, 10, 20, 50, 100, 200, 500, 1000, Inf)
prot_area_palette <- colorBin("YlOrRd", domain = protected_areas_wgs84$n_taxa)

leaflet() %>%
  addTiles() %>%
  addPolygons(data = protected_areas_wgs84 %>%
                filter(SITETYPE == "A"),
              fillColor = ~prot_area_palette(n_taxa),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~n_taxa,
              popup = ~paste(SITECODE, SITENAME, sep = ": "),
              group = "type A") %>%
  addPolygons(data = protected_areas_wgs84 %>%
                filter(SITETYPE == "B"),
              fillColor = ~prot_area_palette(n_taxa),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~n_taxa,
              popup = ~paste(SITECODE, SITENAME, sep = ": "),
              group = "type B") %>%
  addPolygons(data = protected_areas_wgs84 %>%
                filter(SITETYPE == "C"),
              fillColor = ~prot_area_palette(n_taxa),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~n_taxa,
              popup = ~paste(SITECODE, SITENAME, sep = ": "),
              group = "type C") %>%
  addLegend(data = protected_areas_wgs84,
            pal = prot_area_palette,
            values = ~n_taxa,
            opacity = 0.7,
            title = "Number of alien taxa",
            position = "bottomleft") %>%
  addLayersControl(
    overlayGroups = c("type A", "type B", "type C"),
    options = layersControlOptions(collapsed = FALSE))
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_species_obs_per_area_concern}
n_species_obs_per_area_concern <-
  occs_prot_areas %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey) %>%
    left_join(protected_areas_metadata,
              by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE) %>%
  filter(year >= begin_year) %>%
  rename(n_obs = n) %>%
  group_by(SITECODE,
           SITENAME,
           SITETYPE,
           flanders,
           wallonia,
           brussels) %>%
  summarize(n_species =  length(unique(speciesKey)),
            n_obs = sum(n_obs)) %>%
  ungroup() %>%
  arrange(desc(n_species), desc(n_obs)) %>%
  select(SITECODE,
         SITENAME,
         SITETYPE,
         n_species,
         n_obs,
         everything())
n_species_obs_per_area_concern
```

Map of the protected areas with number of species included in the concern list:

```{r leaflet_n_species_in_prot_aras_union_list, echo = FALSE}
# Transform to wgs84
protected_areas_wgs84_concern <- 
  protected_areas %>%
          st_transform(crs = 4326)

protected_areas_wgs84_concern <-
  protected_areas_wgs84_concern %>%
  merge(n_species_obs_per_area_concern %>%
              select(SITECODE, n_species, n_obs),
        all = TRUE)

# bins <- c(0, 10, 20, 50, 100, 200, 500, 1000, Inf)
prot_area_palette_concern <- colorBin(
  "YlOrRd",
  domain = protected_areas_wgs84_concern$n_species
)

leaflet() %>%
  addTiles() %>%
  addPolygons(data = protected_areas_wgs84_concern %>%
                filter(SITETYPE == "A"),
              fillColor = ~prot_area_palette_concern(n_species),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~n_species,
              popup = ~paste(SITECODE, SITENAME, sep = ": "),
              group = "type A") %>%
  addPolygons(data = protected_areas_wgs84_concern %>%
                filter(SITETYPE == "B"),
              fillColor = ~prot_area_palette_concern(n_species),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~n_species,
              popup = ~paste(SITECODE, SITENAME, sep = ": "),
              group = "type B") %>%
  addPolygons(data = protected_areas_wgs84_concern %>%
                filter(SITETYPE == "C"),
              fillColor = ~prot_area_palette_concern(n_species),
              fillOpacity = 0.7,
              color = "black",
              weight = 0.5,
              opacity = 1,
              label = ~n_species,
              popup = ~paste(SITECODE, SITENAME, sep = ": "),
              group = "type C") %>%
  addLegend(data = protected_areas_wgs84_concern,
            pal = prot_area_palette_concern,
            values = ~n_species,
            opacity = 0.7,
            title = "Number of Union list alien species",
            position = "bottomleft") %>%
  addLayersControl(
    overlayGroups = c("type A", "type B", "type C"),
    options = layersControlOptions(collapsed = FALSE))
```

### Flanders

Number of alien taxa and number of observations of alien taxa in each Flemish protected area from 2000 up to now:

```{r n_taxa_obs_per_area_flanders}
n_taxa_obs_per_area_flanders <- 
  n_taxa_obs_per_area %>%
  filter(flanders == TRUE)
n_taxa_obs_per_area_flanders
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_species_obs_per_area_concern_flanders}
n_species_obs_per_area_concern_flanders <-
  n_species_obs_per_area_concern %>%
  filter(flanders == TRUE)
n_species_obs_per_area_concern_flanders
```

### Wallonia

Number of alien taxa and number of observations of alien taxa in each protected area of Wallonia from 2000 up to now:

```{r n_taxa_obs_per_area_wallonia}
n_taxa_obs_per_area_wallonia <-
  n_taxa_obs_per_area %>%
  filter(wallonia == TRUE)
n_taxa_obs_per_area_wallonia
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_species_obs_per_area_concern_wallonia}
n_species_obs_per_area_concern_wallonia <- 
  n_species_obs_per_area_concern %>%
  filter(wallonia == TRUE)
n_species_obs_per_area_concern_wallonia
```
### Brussels

Number of alien taxa and number of observations of alien taxa in each protected area of Brussels from 2000 up to now:

```{r n_taxa_obs_per_area_brussels}
n_taxa_obs_per_area_brussels <-
  n_taxa_obs_per_area %>%
  filter(brussels == TRUE)
n_taxa_obs_per_area_brussels
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_species_obs_per_area_concern_brussels}
n_species_obs_per_area_concern_brussels <- 
  n_species_obs_per_area_concern %>%
  filter(brussels == TRUE)
n_species_obs_per_area_concern_brussels
```

## Bird directive areas

We show the results for the bird directive areas (SPA areas), i.e. areas of `SITETYPE` value `A` or `C`.

### Belgium

Number of alien taxa and number of observations of alien taxa in each SPA area from 2000 up to now:

```{r n_taxa_obs_per_area_bird}
n_taxa_obs_per_area_bird <-
  n_taxa_obs_per_area %>%
  filter(SITETYPE %in% c("A", "C"))
n_taxa_obs_per_area_bird
```

Limited to species from the list of Union concern:

```{r n_species_obs_per_area_concern_bird}
n_species_obs_per_area_concern_bird <-
  n_species_obs_per_area_concern %>%
  filter(SITETYPE %in% c("A", "C"))
n_species_obs_per_area_concern_bird
```

We repeat this analysis for each Belgian region separately.

### Flanders

Number of alien taxa and number of observations of alien taxa in each Flemish SPA area from 2000 up to now:

```{r n_taxa_obs_per_area_flanders_bird}
n_taxa_obs_per_area_flanders_bird <-
  n_taxa_obs_per_area_flanders %>%
  filter(SITETYPE %in% c("A","C"))
n_taxa_obs_per_area_flanders_bird
```

Limited to species from the list of Union concern:

```{r n_species_obs_per_area_concern_flanders_bird}
n_species_obs_per_area_concern_flanders_bird <-
  n_species_obs_per_area_concern_flanders %>%
  filter(SITETYPE %in% c("A","C"))
n_species_obs_per_area_concern_flanders_bird
```

### Wallonia

Number of alien taxa and number of observations of alien taxa in each SPA area of Wallonia from 2000 up to now:

```{r n_taxa_obs_per_area_wallonia_bird}
n_taxa_obs_per_area_wallonia_bird <-
  n_taxa_obs_per_area_wallonia %>%
  filter(SITETYPE %in% c("A","C"))
n_taxa_obs_per_area_wallonia_bird
```

Limited to species from the list of Union concern:

```{r n_species_obs_per_area_concern_wallonia_bird}
n_species_obs_per_area_concern_wallonia_bird <-
  n_species_obs_per_area_concern_wallonia %>%
  filter(SITETYPE %in% c("A","C"))
n_species_obs_per_area_concern_wallonia_bird
```

### Brussels

Number of alien taxa and number of observations of alien taxa in each SPA area of Brussels from 2000 up to now:

```{r n_taxa_obs_per_area_brussels_bird}
n_taxa_obs_per_area_brussels_bird <-
  n_taxa_obs_per_area_brussels %>%
  filter(SITETYPE %in% c("A","C"))
n_taxa_obs_per_area_brussels_bird
```

Limited to species from the list of Union concern:

```{r n_species_obs_per_area_concern_brussels_bird}
n_species_obs_per_area_concern_brussels_bird <-
  n_species_obs_per_area_concern_brussels %>%
  filter(SITETYPE %in% c("A","C"))
n_species_obs_per_area_concern_brussels_bird
```

## Habitat directive areas

We show the results for habitat directive areas (SAC areas), i.e. areas of `SITETYPE` value `B` or `C`.

### Belgium

Number of alien taxa and number of observations of alien taxa in each SAC area from 2000 up to now:

```{r n_taxa_obs_per_area_habitat}
n_taxa_obs_per_area_habitat <-
  n_taxa_obs_per_area %>%
  filter(SITETYPE %in% c("B", "C"))
n_taxa_obs_per_area_habitat
```

Limited to species from the list of Union concern:

```{r n_species_obs_per_area_concern_habitat}
n_species_obs_per_area_concern_habitat <-
  n_species_obs_per_area_concern %>%
  filter(SITETYPE %in% c("B","C"))
n_species_obs_per_area_concern_habitat
```
We repeat the same analysis for each of the three Belgian regions.

### Flanders

Number of alien taxa and number of observations of alien taxa in each Flemish SAC area from 2000 up to now:

```{r n_taxa_obs_per_area_flanders_habitat}
n_taxa_obs_per_area_flanders_habitat <-
  n_taxa_obs_per_area_flanders %>%
  filter(SITETYPE %in% c("B","C"))
n_taxa_obs_per_area_flanders_habitat
```

Limited to species from the list of Union concern:

```{r n_species_obs_per_area_concern_flanders_habitat}
n_species_obs_per_area_concern_flanders_habitat <-
  n_species_obs_per_area_concern_flanders %>%
  filter(SITETYPE %in% c("B","C"))
n_species_obs_per_area_concern_flanders_habitat
```

### Wallonia

Number of alien taxa and number of observations of alien taxa in each SAC area of Wallonia from 2000 up to now:

```{r n_taxa_obs_per_area_wallonia_habitat}
n_taxa_obs_per_area_wallonia_habitat <-
  n_taxa_obs_per_area_wallonia %>%
  filter(SITETYPE %in% c("B","C"))
n_taxa_obs_per_area_wallonia_habitat
```

Limited to species from the list of Union concern:

```{r n_species_obs_per_area_concern_wallonia_habitat}
n_species_obs_per_area_concern_wallonia_habitat <-
  n_species_obs_per_area_concern_wallonia %>%
  filter(SITETYPE %in% c("B","C"))
n_species_obs_per_area_concern_wallonia_habitat
```

### Brussels

Number of alien taxa and number of observations of alien taxa in each SAC area of Brussels from 2000 up to now:

```{r n_taxa_obs_per_area_brussels_habitat}
n_taxa_obs_per_area_brussels_habitat <-
  n_taxa_obs_per_area_brussels %>%
  filter(SITETYPE %in% c("B","C"))
n_taxa_obs_per_area_brussels_habitat
```

Limited to species from the list of Union concern:

```{r n_species_obs_per_area_concern_brussels_habitat}
n_species_obs_per_area_concern_brussels_habitat <-
  n_species_obs_per_area_concern_brussels %>%
  filter(SITETYPE %in% c("B","C"))
n_species_obs_per_area_concern_brussels_habitat
```

# Spread of alien taxa in protected areas

Now we reverse the analysis by counting in how many protected areas each alien taxon has been found. If the taxon is present in at least 10 areas, we calculate the mean coverage of the protected areas as well. Such coverage is the mean of the coverage calculated for each year and each area.

Again, we select observations from 2000. In this case we limit up to the observations of the last year to avoid underestimation of the mean coverage.

```{r begin_year_last_year}
begin_year <- 2000
last_year <- lubridate::year(Sys.Date()) - 1
```

We define first a help function:

```{r define_n_areas_mean_coverage}
#' Calculate number of protected areas a taxon is present
#'
#' This function calculates in how many protected areas a taxon is
#' present. If it is present in more than or equal to 10 protected 
#' areas the mean coverage (`mean_coverage`) is also calculated.
#' @param df data.frame. A data.frame.
n_areas_mean_coverage <- function(df) {
  df <- 
    df %>%
    group_by(taxonKey)
  df <- 
    df %>%
    summarize(n_prot_areas = length(unique(SITECODE)),
              mean_coverage = if_else(length(unique(SITECODE)) >= 10,
                                      mean(coverage, na.rm = TRUE),
                                      NA_real_)) %>%
    ungroup()
  
  df <- 
    df %>%
    left_join(taxonomic_metadata, by = "taxonKey")  %>%
    arrange(desc(n_prot_areas),
            desc(mean_coverage),
            scientificName) %>%
    select(scientificName,
           n_prot_areas,
           mean_coverage,
           taxonKey,
           everything())
  return(df)
}
```

## Natura2000 areas

We start by taking into account all type of areas.

### Belgium

Number of protected areas and their mean coverage for each alien species:

```{r n_prot_areas_per_taxon}
n_prot_areas_per_taxon <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_concern}
n_prot_areas_per_species_concern <-
  n_prot_areas_per_taxon %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_concern
```

We proceed now by analyzing the spreading of alien species at regional level.

### Flanders

Number of Flemish protected areas and their mean coverage for each alien taxon:

```{r n_prot_areas_per_taxon_flanders}
n_prot_areas_per_taxon_flanders <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year,
         flanders == TRUE) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon_flanders
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_flanders_concern}
n_prot_areas_per_species_flanders_concern <-
  n_prot_areas_per_taxon_flanders %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_flanders_concern
```

### Wallonia

Number of protected areas in Wallonia and their mean coverage for each alien species:

```{r n_prot_areas_per_taxon_wallonia}
n_prot_areas_per_taxon_wallonia <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year,
         wallonia == TRUE) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon_wallonia
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_wallonia_concern}
n_prot_areas_per_species_wallonia_concern <-
  n_prot_areas_per_taxon_wallonia %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_wallonia_concern
```

### Brussels

Number of protected areas of Brussels and their mean coverage for each alien taxon:

```{r n_prot_areas_per_taxon_brussels}
n_prot_areas_per_taxon_brussels <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year,
         brussels == TRUE) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon_brussels
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_brussels_concern}
n_prot_areas_per_species_brussels_concern <-
  n_prot_areas_per_taxon_brussels %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_brussels_concern
```

## Bird directive areas

We show the results for the bird directive areas (SPA areas), i.e. areas of `SITETYPE` value `A` or `C`.

### Belgium

Number of SPA protected areas and their mean coverage for each alien taxon:

```{r n_prot_areas_per_taxon_bird}
n_prot_areas_per_taxon_bird <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year,
         SITETYPE %in% c("A", "C")) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon_bird
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_concern_bird}
n_prot_areas_per_species_concern_bird <-
  n_prot_areas_per_taxon_bird %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_concern_bird
```

We proceed now by analyzing the spreading of alien species at regional level.

### Flanders

Number of Flemish SPA protected areas and their mean coverage for each alien species:

```{r n_prot_areas_per_taxon_flanders_bird}
n_prot_areas_per_taxon_flanders_bird <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year,
         SITETYPE %in% c("A", "C"),
         flanders == TRUE) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon_flanders_bird
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_concern_flanders_bird}
n_prot_areas_per_species_concern_flanders_bird <-
  n_prot_areas_per_taxon_flanders_bird %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_concern_flanders_bird
```

### Wallonia

Number of SPA protected areas in Wallonia and their mean coverage for each alien taxon:

```{r n_prot_areas_per_taxon_wallonia_bird}
n_prot_areas_per_taxon_wallonia_bird <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year,
         SITETYPE %in% c("A", "C"),
         wallonia == TRUE) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon_wallonia_bird
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_concern_wallonia_bird}
n_prot_areas_per_species_concern_wallonia_bird <-
  n_prot_areas_per_taxon_wallonia_bird %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_concern_wallonia_bird
```

### Brussels

Number of SPA protected areas in Brussels and their mean coverage for each alien taxon:

```{r n_prot_areas_per_taxon_Brussels_bird}
n_prot_areas_per_taxon_brussels_bird <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year,
         SITETYPE %in% c("A", "C"),
         brussels == TRUE) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon_brussels_bird
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_concern_brussels_bird}
n_prot_areas_per_species_concern_brussels_bird <-
  n_prot_areas_per_taxon_brussels_bird %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_concern_brussels_bird
```

## Habitat directive areas

We show the results for the habitat directive areas (SAC areas), i.e. areas of `SITETYPE` value `B` or `C`.

### Belgium

Number of SAC protected areas and their mean coverage for each alien taxon:

```{r n_prot_areas_per_taxon_habitat}
n_prot_areas_per_taxon_habitat <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year,
         SITETYPE %in% c("B", "C")) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon_habitat
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_concern_habitat}
n_prot_areas_per_species_concern_habitat <-
  n_prot_areas_per_taxon_habitat %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_concern_habitat
```

We proceed now by analyzing the spreading of alien taxa at regional level.

### Flanders

Number of Flemish SAC protected areas and their mean coverage for each alien species:

```{r n_prot_areas_per_taxon_flanders_habitat}
n_prot_areas_per_taxon_flanders_habitat <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year,
         SITETYPE %in% c("B", "C"),
         flanders == TRUE) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon_flanders_habitat
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_concern_flanders_habitat}
n_prot_areas_per_species_concern_flanders_habitat <-
  n_prot_areas_per_taxon_flanders_habitat %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_concern_flanders_habitat
```

### Wallonia

Number of SAC protected areas in Wallonia and their mean coverage for each alien taxon:

```{r n_prot_areas_per_taxon_wallonia_habitat}
n_prot_areas_per_taxon_wallonia_habitat <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year,
         SITETYPE %in% c("B", "C"),
         wallonia == TRUE) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon_wallonia_habitat
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_concern_wallonia_habitat}
n_prot_areas_per_species_concern_wallonia_habitat <-
  n_prot_areas_per_taxon_wallonia_habitat %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_concern_wallonia_habitat
```

### Brussels

Number of SAC protected areas in Brussels and their mean coverage for each alien taxon:

```{r n_prot_areas_per_taxon_Brussels_habitat}
n_prot_areas_per_taxon_brussels_habitat <-
  occs_prot_areas %>%
  left_join(protected_areas_metadata,
            by = c("SITECODE", "SITETYPE")) %>%
  filter(is_alien == TRUE,
         year >= begin_year,
         year <= last_year,
         SITETYPE %in% c("B", "C"),
         brussels == TRUE) %>%
  n_areas_mean_coverage()
n_prot_areas_per_taxon_brussels_habitat
```

Limited to species included in the list of invasive alien species of Union concern:

```{r n_prot_areas_per_species_concern_brussels_habitat}
n_prot_areas_per_species_concern_brussels_habitat <-
  n_prot_areas_per_taxon_brussels_habitat %>%
  filter(taxonKey %in% eu_concern_species$backbone_taxonKey)
n_prot_areas_per_species_concern_brussels_habitat
```

# Save data

## Save stage of invasion data

We save the analysis of the stage of the invasion of the protected areas:

```{r save_stage_of_invasion_data}
n_taxa_obs_per_area_dfs <- list(
  # Belgium
  n_taxa_obs_per_area = n_taxa_obs_per_area,
  n_species_obs_per_area_concern = n_species_obs_per_area_concern,
  n_taxa_obs_per_area_bird = n_taxa_obs_per_area_bird,
  n_species_obs_per_area_concern_bird = n_species_obs_per_area_concern_bird,
  n_taxa_obs_per_area_habitat = n_taxa_obs_per_area_habitat,
  n_species_obs_per_area_concern_habitat = 
    n_species_obs_per_area_concern_habitat,
  # Flanders
  n_taxa_obs_per_area_flanders = n_taxa_obs_per_area_flanders,
  n_species_obs_per_area_concern_flanders = 
    n_species_obs_per_area_concern_flanders,
  n_taxa_obs_per_area_flanders_bird = n_taxa_obs_per_area_flanders_bird,
  n_species_obs_per_area_concern_flanders_bird = 
    n_species_obs_per_area_concern_flanders_bird,
  n_taxa_obs_per_area_flanders_habitat = n_taxa_obs_per_area_flanders_habitat,
  n_species_obs_per_area_concern_flanders_habitat = 
    n_species_obs_per_area_concern_flanders_habitat,
  # Wallonia
  n_taxa_obs_per_area_wallonia = n_taxa_obs_per_area_wallonia,
  n_species_obs_per_area_concern_wallonia = 
    n_species_obs_per_area_concern_wallonia,
  n_taxa_obs_per_area_wallonia_bird = n_taxa_obs_per_area_wallonia_bird,
  n_species_obs_per_area_concern_wallonia_bird = 
    n_species_obs_per_area_concern_wallonia_bird,
  n_taxa_obs_per_area_wallonia_habitat = n_taxa_obs_per_area_wallonia_habitat,
  n_species_obs_per_area_concern_wallonia_habitat = 
    n_species_obs_per_area_concern_wallonia_habitat,
  # Brussels
  n_taxa_obs_per_area_brussels = n_taxa_obs_per_area_brussels,
  n_species_obs_per_area_concern_brussels = 
    n_species_obs_per_area_concern_brussels,
  n_taxa_obs_per_area_brussels_bird = n_taxa_obs_per_area_brussels_bird,
  n_species_obs_per_area_concern_brussels_bird = 
    n_species_obs_per_area_concern_brussels_bird,
  n_taxa_obs_per_area_brussels_habitat = n_taxa_obs_per_area_brussels_habitat,
  n_species_obs_per_area_concern_brussels_habitat = 
    n_species_obs_per_area_concern_brussels_habitat
)
map2(n_taxa_obs_per_area_dfs,
     names(n_taxa_obs_per_area_dfs),
     function(df, filename) {
       write_csv(df,
                 here::here("data",
                            "output",
                            "protected_areas_analysis",
                            paste0(filename, ".csv")),
                 na = "")
     })
```

## Save spreading data

We save the analysis of the spreading of alien taxa over protected areas:

```{r save_spreading_data}
n_prot_areas_per_species_dfs <-
  list(
    # Belgium
    n_prot_areas_per_taxon = n_prot_areas_per_taxon,
    n_prot_areas_per_species_concern = n_prot_areas_per_species_concern,
    n_prot_areas_per_taxon_bird = n_prot_areas_per_taxon_bird,
    n_prot_areas_per_species_concern_bird = 
      n_prot_areas_per_species_concern_bird,
    n_prot_areas_per_taxon_habitat = n_prot_areas_per_taxon_habitat,
    n_prot_areas_per_species_concern_habitat = 
      n_prot_areas_per_species_concern_habitat,
    # Flanders
    n_prot_areas_per_taxon_flanders = n_prot_areas_per_taxon_flanders,
    n_prot_areas_per_species_flanders_concern = 
      n_prot_areas_per_species_flanders_concern,
    n_prot_areas_per_taxon_flanders_bird = 
      n_prot_areas_per_taxon_flanders_bird,
    n_prot_areas_per_species_concern_flanders_bird = 
      n_prot_areas_per_species_concern_flanders_bird,
    n_prot_areas_per_taxon_flanders_habitat = 
      n_prot_areas_per_taxon_flanders_habitat,
    n_prot_areas_per_species_concern_flanders_habitat = 
      n_prot_areas_per_species_concern_flanders_habitat,
    # Wallonia
    n_prot_areas_per_taxon_wallonia = n_prot_areas_per_taxon_wallonia,
    n_prot_areas_per_species_wallonia_concern = 
      n_prot_areas_per_species_wallonia_concern,
    n_prot_areas_per_taxon_wallonia_bird = 
      n_prot_areas_per_taxon_wallonia_bird,
    n_prot_areas_per_species_concern_wallonia_bird = 
      n_prot_areas_per_species_concern_wallonia_bird,
    n_prot_areas_per_taxon_wallonia_habitat = 
      n_prot_areas_per_taxon_wallonia_habitat,
    n_prot_areas_per_species_concern_wallonia_habitat = 
      n_prot_areas_per_species_concern_wallonia_habitat,
    # Brussels
    n_prot_areas_per_taxon_brussels = n_prot_areas_per_taxon_brussels,
    n_prot_areas_per_species_brussels_concern = 
      n_prot_areas_per_species_brussels_concern,
    n_prot_areas_per_taxon_brussels_bird = 
      n_prot_areas_per_taxon_brussels_bird,
    n_prot_areas_per_species_concern_brussels_bird = 
      n_prot_areas_per_species_concern_brussels_bird,
    n_prot_areas_per_taxon_brussels_habitat = 
      n_prot_areas_per_taxon_brussels_habitat,
    n_prot_areas_per_species_concern_brussels_habitat = 
      n_prot_areas_per_species_concern_brussels_habitat
  )
# Save data
map2(n_prot_areas_per_species_dfs,
     names(n_prot_areas_per_species_dfs),
     function(df, filename) {
       write_csv(df,
                 here::here("data",
                            "output",
                            "protected_areas_analysis",
                            paste0(filename, ".csv")),
                 na = "")
     })
```
